reorganization of an OpenCV dependency
