+++ gpu/kinfu/src/kinfu.cpp
@@ -51,7 +51,7 @@
 #ifdef HAVE_OPENCV
   #include <opencv2/opencv.hpp>
   #include <opencv2/gpu/gpu.hpp>
-  #include <pcl/gpu/utils/timers_opencv.hpp>
+  #include <pcl/impl/timers_opencv.hpp>
 #endif
 
 using namespace std;
