+++ gpu/kinfu/tools/kinfu_app_sim.cpp
@@ -91,9 +91,9 @@
 #ifdef HAVE_OPENCV  
   #include <opencv2/highgui/highgui.hpp>
   #include <opencv2/imgproc/imgproc.hpp>
-  #include <pcl/gpu/utils/timers_opencv.hpp>
+  #include <pcl/impl/timers_opencv.hpp>
 //#include "video_recorder.h"
-typedef pcl::gpu::ScopeTimerCV ScopeTimeT;
+typedef pcl::ScopeTimerCV ScopeTimeT;
 #else
   typedef pcl::ScopeTime ScopeTimeT;
 #endif
